// src/types/data.ts
export interface TypeData {
    locale: string;
    zIndex: number;
    width: number;
    height: number;
    backgroundImage: string;
    scheduleStart: Date;
    scheduleDateFormat: string;
    scheduleWeekFormat: string;
    scheduleDateDesign: string;
    scheduleWeekDesign: string;
    scheduleContentDesign: string;
}